src/machine_cnc/requirements.txt
Flask==3.1.0
kafka-python==2.1.5